Move UI related classes to their own package (core.ui)
