+++ src/main/java/org/istic/synthlab/components/vcoa/Controller.java
@@ -3,7 +3,7 @@ package org.istic.synthlab.components.vcoa;
 import javafx.fxml.FXML;
 import javafx.scene.shape.Circle;
 import org.istic.synthlab.core.AbstractController;
-import org.istic.synthlab.core.IHMConnectionManager;
+import org.istic.synthlab.core.ui.ConnectionManager;
 
 import java.net.URL;
 import java.util.ResourceBundle;
@@ -15,13 +15,13 @@ public class Controller extends AbstractController {
     @FXML
     Circle input;
     private static int numInstance = 0;
-    private Vcoa composantVcoa = new Vcoa("VCOA"+numInstance++);;
+    private Vcoa composantVcoa = new Vcoa("VCOA"+numInstance++);
 
     public void initialize(URL location, ResourceBundle resources) {
     }
 
     @FXML
     void connectIn(){
-        IHMConnectionManager.makeOrigin(composantVcoa.getOutput());
+        ConnectionManager.makeOrigin(composantVcoa.getOutput());
     }
 }
